+++ src/plugins/cache/operations/no-operation.js
@@ -1,5 +1,5 @@
 import {passThrough} from 'ladda-fp';
-import {invalidate} from './query-cache';
+import {invalidate} from '../query-cache';
 
 export function decorateNoOperation(c, es, qc, e, aFn) {
   return (...args) => {
